As far as I can tell, the `map_repeat` function just does exactly what Julia v0.6-style broadcasting over tuple arguments does. 

I'm currently tuning up the benchmark suite to try to verify that this has no performance penalty. 

Fixes #196