@rdeits 2018-03-01
Here are the benchmark changes, although I'm not sure how to interpret them:

```julia
julia> leaves(improvements(judge(minimum(pr), minimum(master))))
1-element Array{Any,1}:
 (Any["bsplines", "quadratic", "2d_Quadratic(Line())_OnCell_use"], TrialJudgement(-13.19% => improvement))

julia> leaves(regressions(judge(minimum(pr), minimum(master))))
3-element Array{Any,1}:
 (Any["bsplines", "constant", "1d_Constant()_OnCell_use"], TrialJudgement(+8.24% => regression))         
 (Any["bsplines", "linear", "3d_Linear()_OnGrid_use"], TrialJudgement(+9.65% => regression))             
 (Any["bsplines", "quadratic", "3d_Quadratic(Flat())_OnGrid_use"], TrialJudgement(+10.27% => regression))
```