@rdeits 2018-03-01
Targeted benchmarking of just the `indices()` function before and after this PR shows no change:

```julia
using Interpolations
using BenchmarkTools

A_x1 = 1:.1:10
A_x2 = 1:.5:20
f(x1, x2) = log(x1+x2)
A = [f(x1,x2) for x1 in A_x1, x2 in A_x2];
itp = interpolate(A, BSpline(Cubic(Line())), OnGrid());

@benchmark indices($itp)
```

Master:

```
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.396 ns (0.00% GC)
  median time:      2.412 ns (0.00% GC)
  mean time:        2.459 ns (0.00% GC)
  maximum time:     16.628 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```

This PR:

```
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.408 ns (0.00% GC)
  median time:      2.419 ns (0.00% GC)
  mean time:        2.451 ns (0.00% GC)
  maximum time:     13.812 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```